+++ Packages~/YooAsset/Runtime/AssetSystem/Loader/RawBundleFileLoader.cs
@@ -16,7 +16,6 @@ namespace YooAsset
 		}
 
 		private ESteps _steps = ESteps.None;
-		private bool _isShowWaitForAsyncError = false;
 		private DownloaderBase _unpacker;
 		private DownloaderBase _downloader;
 
@@ -42,7 +41,7 @@ namespace YooAsset
 				}
 				else if (MainBundleInfo.LoadMode == BundleInfo.ELoadMode.LoadFromStreaming)
 				{
-#if UNITY_ANDROID || UNITY_WEBGL
+#if UNITY_ANDROID
 					_steps = ESteps.Unpack;
 					FileLoadPath = MainBundleInfo.Bundle.CachedDataFilePath;
 #else
@@ -125,13 +124,14 @@ namespace YooAsset
 				DownloadProgress = 1f;
 				DownloadedBytes = (ulong)MainBundleInfo.Bundle.FileSize;
 
-				_steps = ESteps.Done;
 				if (File.Exists(FileLoadPath))
 				{
+					_steps = ESteps.Done;
 					Status = EStatus.Succeed;
 				}
 				else
 				{
+					_steps = ESteps.Done;
 					Status = EStatus.Failed;
 					LastError = $"Raw file not found : {FileLoadPath}";
 				}
@@ -149,20 +149,24 @@ namespace YooAsset
 				// 文件解压
 				if (_unpacker != null)
 				{
-					_unpacker.Update();
 					if (_unpacker.IsDone() == false)
+					{
+						_unpacker.WaitForAsyncComplete = true;
+						_unpacker.Update();
 						continue;
+					}
 				}
 
 				// 保险机制
-				// 注意：如果需要从WEB端下载资源，可能会触发保险机制！
+				// 注意：如果需要从远端下载资源，可能会触发保险机制！
 				frame--;
 				if (frame == 0)
 				{
-					if (_isShowWaitForAsyncError == false)
+					if (IsDone() == false)
 					{
-						_isShowWaitForAsyncError = true;
-						YooLogger.Error($"WaitForAsyncComplete failed ! Try load bundle : {MainBundleInfo.Bundle.BundleName} from remote with sync load method !");
+						Status = EStatus.Failed;
+						LastError = $"WaitForAsyncComplete failed ! Try load bundle : {MainBundleInfo.Bundle.BundleName} from remote with sync load method !";
+						YooLogger.Error(LastError);
 					}
 					break;
 				}