+++ Packages~/YooAsset/Runtime/DownloadSystem/DownloadSystem.cs
@@ -2,10 +2,16 @@
 using System.IO;
 using System.Collections;
 using System.Collections.Generic;
+using System.Threading;
 using UnityEngine.Networking;
 
 namespace YooAsset
 {
+	/// <summary>
+	/// 自定义下载器的请求委托
+	/// </summary>
+	public delegate UnityWebRequest DownloadRequestDelegate(string url);
+
 	/// <summary>
 	/// 1. 保证每一时刻资源文件只存在一个下载器
 	/// 2. 保证下载器下载完成后立刻验证并缓存
@@ -16,11 +22,15 @@ namespace YooAsset
 		private static readonly Dictionary<string, DownloaderBase> _downloaderDic = new Dictionary<string, DownloaderBase>();
 		private static readonly List<string> _removeList = new List<string>(100);
 
+		/// <summary>
+		/// 自定义下载器的请求委托
+		/// </summary>
+		public static DownloadRequestDelegate RequestDelegate = null;
 
 		/// <summary>
 		/// 自定义的证书认证实例
 		/// </summary>
-		public static CertificateHandler CertificateHandlerInstance;
+		public static CertificateHandler CertificateHandlerInstance = null;
 
 		/// <summary>
 		/// 启用断点续传功能文件的最小字节数
@@ -32,8 +42,16 @@ namespace YooAsset
 		/// </summary>
 		public static List<long> ClearFileResponseCodes { set; get; }
 
+
 		/// <summary>
-		/// 更新所有下载器
+		/// 初始化下载器
+		/// </summary>
+		public static void Initialize()
+		{
+		}
+
+		/// <summary>
+		/// 更新下载器
 		/// </summary>
 		public static void Update()
 		{
@@ -66,7 +84,11 @@ namespace YooAsset
 			}
 			_downloaderDic.Clear();
 			_removeList.Clear();
+
+			RequestDelegate = null;
+			CertificateHandlerInstance = null;
 			BreakpointResumeFileSize = int.MaxValue;
+			ClearFileResponseCodes = null;
 		}
 
 
@@ -99,6 +121,18 @@ namespace YooAsset
 				_downloaderDic.Add(bundleInfo.Bundle.CachedDataFilePath, newDownloader);
 				return newDownloader;
 			}
+		}
+
+		/// <summary>
+		/// 创建一个新的网络请求
+		/// </summary>
+		public static UnityWebRequest NewRequest(string requestURL)
+		{
+			if (RequestDelegate != null)
+				return RequestDelegate.Invoke(requestURL);
+
+			var request = new UnityWebRequest(requestURL, UnityWebRequest.kHttpVerbGET);
+			return request;
 		}
 
 		/// <summary>