+++ Packages~/YooAsset/Runtime/DownloadSystem/Downloader/FileDownloader.cs
@@ -13,6 +13,8 @@ namespace YooAsset
 		private readonly string _tempFilePath;
 		private UnityWebRequest _webRequest = null;
 		private DownloadHandlerFileRange _downloadHandle = null;
+		private VerifyTempFileOperation _checkFileOp = null;
+		private VerifyTempFileOperation _verifyFileOp = null;
 
 		// 重置变量
 		private bool _isAbort = false;
@@ -34,17 +36,31 @@ namespace YooAsset
 			if (IsDone())
 				return;
 
-			// 检测本地临时文件
+			// 检测临时文件
 			if (_steps == ESteps.CheckTempFile)
 			{
-				var verifyResult = CacheSystem.VerifyingTempFile(_bundleInfo.Bundle, EVerifyLevel.High);
-				if (verifyResult == EVerifyResult.Succeed)
+				VerifyTempElement element = new VerifyTempElement(_bundleInfo.Bundle.TempDataFilePath, _bundleInfo.Bundle.FileCRC, _bundleInfo.Bundle.FileSize);
+				_checkFileOp = VerifyTempFileOperation.CreateOperation(element);
+				OperationSystem.StartOperation(_checkFileOp);
+				_steps = ESteps.WaitingCheckTempFile;
+			}
+
+			// 等待检测结果
+			if (_steps == ESteps.WaitingCheckTempFile)
+			{
+				if (WaitForAsyncComplete)
+					_checkFileOp.Update();
+
+				if (_checkFileOp.IsDone == false)
+					return;
+
+				if (_checkFileOp.Status == EOperationStatus.Succeed)
 				{
 					_steps = ESteps.CachingFile;
 				}
 				else
 				{
-					if (verifyResult == EVerifyResult.FileOverflow)
+					if (_checkFileOp.VerifyResult == EVerifyResult.FileOverflow)
 					{
 						if (File.Exists(_tempFilePath))
 							File.Delete(_tempFilePath);
@@ -80,7 +96,7 @@ namespace YooAsset
 				if (File.Exists(_tempFilePath))
 					File.Delete(_tempFilePath);
 
-				_webRequest = new UnityWebRequest(_requestURL, UnityWebRequest.kHttpVerbGET);
+				_webRequest = DownloadSystem.NewRequest(_requestURL);
 				DownloadHandlerFile handler = new DownloadHandlerFile(_tempFilePath);
 				handler.removeFileOnAbort = true;
 				_webRequest.downloadHandler = handler;
@@ -109,11 +125,11 @@ namespace YooAsset
 				}
 
 #if UNITY_2019_4_OR_NEWER
-				_webRequest = new UnityWebRequest(_requestURL, UnityWebRequest.kHttpVerbGET);
+				_webRequest = DownloadSystem.NewRequest(_requestURL);
 				var handler = new DownloadHandlerFile(_tempFilePath, true);
 				handler.removeFileOnAbort = false;
 #else
-				_webRequest = new UnityWebRequest(_requestURL, UnityWebRequest.kHttpVerbGET);
+				_webRequest = DownloadSystem.NewRequest(_requestURL);
 				var handler = new DownloadHandlerFileRange(_tempFilePath, _bundleInfo.Bundle.FileSize, _webRequest);
 				_downloadHandle = handler;
 #endif
@@ -188,7 +204,7 @@ namespace YooAsset
 				}
 				else
 				{
-					_steps = ESteps.VerifyingFile;
+					_steps = ESteps.VerifyTempFile;
 				}
 
 				// 释放下载器
@@ -196,21 +212,33 @@ namespace YooAsset
 			}
 
 			// 验证下载文件
-			if (_steps == ESteps.VerifyingFile)
+			if (_steps == ESteps.VerifyTempFile)
 			{
-				var verifyResult = CacheSystem.VerifyingTempFile(_bundleInfo.Bundle, EVerifyLevel.High);
-				if (verifyResult == EVerifyResult.Succeed)
+				VerifyTempElement element = new VerifyTempElement(_bundleInfo.Bundle.TempDataFilePath, _bundleInfo.Bundle.FileCRC, _bundleInfo.Bundle.FileSize);
+				_verifyFileOp = VerifyTempFileOperation.CreateOperation(element);
+				OperationSystem.StartOperation(_verifyFileOp);
+				_steps = ESteps.WaitingVerifyTempFile;
+			}
+
+			// 等待验证完成
+			if (_steps == ESteps.WaitingVerifyTempFile)
+			{
+				if (WaitForAsyncComplete)
+					_verifyFileOp.Update();
+
+				if (_verifyFileOp.IsDone == false)
+					return;
+
+				if (_verifyFileOp.Status == EOperationStatus.Succeed)
 				{
 					_steps = ESteps.CachingFile;
 				}
 				else
 				{
-					_lastError = $"Failed to verifying file : {_bundleInfo.Bundle.FileName}, ErrorCode : {verifyResult}";
-
-					// 注意：验证失败后删除文件
 					if (File.Exists(_tempFilePath))
 						File.Delete(_tempFilePath);
 
+					_lastError = _verifyFileOp.Error;
 					_steps = ESteps.TryAgain;
 				}
 			}
@@ -220,22 +248,24 @@ namespace YooAsset
 			{
 				try
 				{
-					string destFilePath = _bundleInfo.Bundle.CachedDataFilePath;
-					if (File.Exists(destFilePath))
-						File.Delete(destFilePath);
+					string infoFilePath = _bundleInfo.Bundle.CachedInfoFilePath;
+					string dataFilePath = _bundleInfo.Bundle.CachedDataFilePath;
+					string dataFileCRC = _bundleInfo.Bundle.FileCRC;
+					long dataFileSize = _bundleInfo.Bundle.FileSize;
+
+					if (File.Exists(infoFilePath))
+						File.Delete(infoFilePath);
+					if (File.Exists(dataFilePath))
+						File.Delete(dataFilePath);
 
 					FileInfo fileInfo = new FileInfo(_tempFilePath);
-					fileInfo.MoveTo(destFilePath);
+					fileInfo.MoveTo(dataFilePath);
 
 					// 写入信息文件记录验证数据
-					CacheFileInfo cacheInfo = new CacheFileInfo();
-					cacheInfo.FileCRC = _bundleInfo.Bundle.FileCRC;
-					cacheInfo.FileSize = _bundleInfo.Bundle.FileSize;
-					string jsonContent = UnityEngine.JsonUtility.ToJson(cacheInfo);
-					FileUtility.CreateFile(_bundleInfo.Bundle.CachedInfoFilePath, jsonContent);
+					CacheFileInfo.WriteInfoToFile(infoFilePath, dataFileCRC, dataFileSize);
 
 					// 记录缓存文件
-					var wrapper = new PackageCache.RecordWrapper(_bundleInfo.Bundle.CachedInfoFilePath, _bundleInfo.Bundle.CachedDataFilePath, _bundleInfo.Bundle.FileCRC, _bundleInfo.Bundle.FileSize);
+					var wrapper = new PackageCache.RecordWrapper(infoFilePath, dataFilePath, dataFileCRC, dataFileSize);
 					CacheSystem.RecordFile(_bundleInfo.Bundle.PackageName, _bundleInfo.Bundle.CacheGUID, wrapper);
 
 					_lastError = string.Empty;