+++ Tools~/Common/Attribute/Script/Assembly/DllAutomation/DllAutomation.cs
@@ -1,13 +1,17 @@
-﻿using System;
+﻿/*|==========|*|
+|*|Author:   |*| -> SAM
+|*|Date:     |*| -> 2023-06-12
+|*|==========|*/
+
+using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Reflection;
-using System.Threading.Tasks;
+using System.Text;
 
 /// <summary>
 /// DLL自动化处理
 /// </summary>
-public static class DllAutomation
+public static partial class DllAutomation
 {
     private static Dictionary<int, List<(DllAutomationAttribute, IDllAutomation)>> Dic;
 
@@ -45,6 +49,7 @@ public static class DllAutomation
                 }
             }
         }
+
         Order.Sort((x, y) =>
         {
             if (x < y) return -1;
@@ -79,6 +84,7 @@ public static class DllAutomation
                 }
             }
         }
+
         Order.Sort((x, y) =>
         {
             if (x < y) return -1;
@@ -87,149 +93,22 @@ public static class DllAutomation
         });
     }
 
-    #region Install
-
     /// <summary>
-    /// 实现计时器
+    /// 输出信息
     /// </summary>
-    public static async Task InstallTimer()
+    public static string OutPut()
     {
-        var watch = new Stopwatch();
-        var allDic = new Dictionary<string, TimeSpan>();
-        var sort = new Dictionary<int, string>();
-        var num = 0;
-        var allTimer = TimeSpan.Zero;
-
-        for (int i = 0; i < Order.Count; i++)
+        var str = new StringBuilder();
+        foreach (var index in Order)
         {
-            foreach (var system in Dic[Order[i]])
+            str.AppendLine("Order : " + index + " -> ");
+            foreach (var (attr, automation) in Dic[index])
             {
-                var alltimelong = watch.Elapsed;
-                try
-                {
-                    watch.Start();
-                    await system.Item2.Install();
-                    watch.Stop();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine("Automation Installation DLL : {0} -> Error : {1}", system.Item1.Name, ex);
-                    watch.Stop();
-                }
-
-                var timer = watch.Elapsed - alltimelong;
-                var title = string.Format("{0} {1}", num, system.Item1.Name);
-                allDic.Add(title, timer);
-                sort.Add(num, title);
-                allTimer += timer;
-                num++;
+                str.AppendLine(string.Format("Priority : {0} -> {1} -> {2}",
+                    attr.Priority, attr.Name, attr.SystemType.FullName));
             }
         }
 
-        Console.WriteLine("Automation Installation DLL Time Total -> <color=#E47833><b>{0:G}</b></color>", allTimer);
-        for (int i = 0; i < num; i++)
-        {
-            string key = sort[i];
-            TimeSpan timer = allDic[key];
-            var timeinfo = string.Format("[<color=#E47833><b>{1}% | {0:G}</b></color>]", timer, (timer.Ticks / (double)allTimer.Ticks * 100).ToString("00.0000")).PadRight(30);
-            Console.WriteLine("Automation Installation DLL Time : {0} -> {1}", timeinfo, key);
-        }
-
+        return str.ToString();
     }
-
-    /// <summary>
-    /// 实现计时器
-    /// </summary>
-    public static async Task Install()
-    {
-        for (int i = 0; i < Order.Count; i++)
-        {
-            foreach (var system in Dic[Order[i]])
-            {
-                try
-                {
-                    await system.Item2.Install();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine("Automation Installation DLL: {0} -> Error : {1}", system.Item1.Name, ex);
-                }
-            }
-        }
-    }
-
-    #endregion
-
-    #region UnInstall
-
-    /// <summary>
-    /// 卸载 实现计时器
-    /// </summary>
-    public static async Task UnInstallTimer()
-    {
-        var watch = new Stopwatch();
-        var allDic = new Dictionary<string, TimeSpan>();
-        var sort = new Dictionary<int, string>();
-        var num = 0;
-        var allTimer = TimeSpan.Zero;
-
-        for (int i = Order.Count - 1; i >= 0; i--)
-        {
-            foreach (var system in Dic[Order[i]])
-            {
-                var alltimelong = watch.Elapsed;
-                try
-                {
-                    watch.Start();
-                    await system.Item2.UnInstall();
-                    watch.Stop();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine("Automation UnInstallation DLL : {0} -> Error : {1}", system.Item1.Name, ex);
-                    watch.Stop();
-                }
-
-                var timer = watch.Elapsed - alltimelong;
-                var title = string.Format("{0} {1}", num, system.Item1.Name);
-                allDic.Add(title, timer);
-                sort.Add(num, title);
-                allTimer += timer;
-                num++;
-            }
-        }
-
-        Console.WriteLine("Automation UnInstallation DLL Time Total -> <color=#E47833><b>{0:G}</b></color>", allTimer);
-        for (int i = 0; i < num; i++)
-        {
-            string key = sort[i];
-            TimeSpan timer = allDic[key];
-            var timeinfo = string.Format("[<color=#E47833><b>{1}% | {0:G}</b></color>]", timer, (timer.Ticks / (double)allTimer.Ticks * 100).ToString("00.0000")).PadRight(30);
-            Console.WriteLine("Automation UnInstallation DLL Time : {0} -> {1}", timeinfo, key);
-        }
-
-    }
-
-    /// <summary>
-    /// 实现计时器
-    /// </summary>
-    public static async Task Uninstall()
-    {
-        for (int i = Order.Count - 1; i >= 0; i--)
-        {
-            foreach (var system in Dic[Order[i]])
-            {
-                try
-                {
-                    await system.Item2.UnInstall();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine("Automation UnInstallation DLL: {0} -> Error : {1}", system.Item1.Name, ex);
-                }
-            }
-        }
-    }
-
-    #endregion
 }