+++ Tools~/Common/Attribute/Script/Class/SystemAutomation/SystemAutomation.cs
@@ -1,13 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Reflection;
-using System.Threading.Tasks;
+using System.Text;
 
 /// <summary>
 /// 系统自动化注册
 /// </summary>
-public static class SystemAutomation
+public static partial class SystemAutomation
 {
     private static Dictionary<int, List<(SystemAutomationAttribute, ISystemAutomation)>> Dic;
 
@@ -45,6 +44,7 @@ public static class SystemAutomation
                 }
             }
         }
+
         Order.Sort((x, y) =>
         {
             if (x < y) return -1;
@@ -62,23 +62,24 @@ public static class SystemAutomation
         Order = new List<int>();
         foreach (var assembly in assemblies)
         {
-            foreach (var type in assembly.GetCustomAttributes<SystemAutomationAttribute>())
+            foreach (var attribute in assembly.GetCustomAttributes<SystemAutomationAttribute>())
             {
-                if (type.SystemType.GetInterface(InterfaceName) != null)
+                if (attribute.SystemType.GetInterface(InterfaceName) != null)
                 {
-                    if (Dic.ContainsKey(type.Priority))
+                    if (Dic.ContainsKey(attribute.Priority))
                     {
-                        Dic[type.Priority].Add((type, (ISystemAutomation)Activator.CreateInstance(type.SystemType)));
+                        Dic[attribute.Priority].Add((attribute, (ISystemAutomation)Activator.CreateInstance(attribute.SystemType)));
                     }
                     else
                     {
-                        Order.Add(type.Priority);
-                        Dic.Add(type.Priority, new List<(SystemAutomationAttribute, ISystemAutomation)>());
-                        Dic[type.Priority].Add((type, (ISystemAutomation)Activator.CreateInstance(type.SystemType)));
+                        Order.Add(attribute.Priority);
+                        Dic.Add(attribute.Priority, new List<(SystemAutomationAttribute, ISystemAutomation)>());
+                        Dic[attribute.Priority].Add((attribute, (ISystemAutomation)Activator.CreateInstance(attribute.SystemType)));
                     }
                 }
             }
         }
+
         Order.Sort((x, y) =>
         {
             if (x < y) return -1;
@@ -87,150 +88,22 @@ public static class SystemAutomation
         });
     }
 
-    #region Install
-
     /// <summary>
-    /// 实现计时器
+    /// 输出信息
     /// </summary>
-    public static async Task InstallTimer()
+    public static string OutPut()
     {
-        var watch = new Stopwatch();
-        var allDic = new Dictionary<string, TimeSpan>();
-        var sort = new Dictionary<int, string>();
-        var num = 0;
-        var allTimer = TimeSpan.Zero;
-
-        for (int i = 0; i < Order.Count; i++)
+        var str = new StringBuilder();
+        foreach (var index in Order)
         {
-            foreach (var system in Dic[Order[i]])
+            str.AppendLine("Order : " + index + " -> ");
+            foreach (var (attribute, automation) in Dic[index])
             {
-                var alltimelong = watch.Elapsed;
-                try
-                {
-                    watch.Start();
-                    await system.Item2.Install();
-                    watch.Stop();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine("Automation Installation System : {0} -> Error : {1}", system.Item1.Name, ex);
-                    watch.Stop();
-                }
-
-                var timer = watch.Elapsed - alltimelong;
-                var title = string.Format("{0} {1}", num, system.Item1.Name);
-                allDic.Add(title, timer);
-                sort.Add(num, title);
-                allTimer += timer;
-                num++;
+                str.AppendLine(string.Format("Priority : {0} -> {1} -> {2}",
+                    attribute.Priority, attribute.Name, attribute.SystemType.FullName));
             }
         }
 
-        Console.WriteLine("Automation Installation System Time Total -> <color=#E47833><b>{0:G}</b></color>", allTimer);
-        for (int i = 0; i < num; i++)
-        {
-            string key = sort[i];
-            TimeSpan timer = allDic[key];
-            var timeinfo = string.Format("[<color=#E47833><b>{1}% | {0:G}</b></color>]", timer, (timer.Ticks / (double)allTimer.Ticks * 100).ToString("00.0000")).PadRight(30);
-            Console.WriteLine("Automation Installation System Time : {0} -> {1}", timeinfo, key);
-        }
-
+        return str.ToString();
     }
-
-    /// <summary>
-    /// 实现计时器
-    /// </summary>
-    public static async Task Install()
-    {
-        for (int i = 0; i < Order.Count; i++)
-        {
-            foreach (var system in Dic[Order[i]])
-            {
-                try
-                {
-                    await system.Item2.Install();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine("Automation Installation System : {0} -> Error : {1}", system.Item1.Name, ex);
-                }
-
-            }
-        }
-    }
-
-    #endregion
-
-    #region Uninstall
-
-    /// <summary>
-    /// 卸载 
-    /// </summary>
-    public static async Task Uninstall()
-    {
-        for (int i = Order.Count - 1; i >= 0; i--)
-        {
-            foreach (var system in Dic[Order[i]])
-            {
-                try
-                {
-                    await system.Item2.UnInstall();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine("Automation UnInstallation System : {0} -> Error : {1}", system.Item1.Name, ex);
-                }
-            }
-        }
-    }
-
-    /// <summary>
-    /// 卸载 实现计时器
-    /// </summary>
-    public static async Task UnInstallTimer()
-    {
-        var watch = new Stopwatch();
-        var allDic = new Dictionary<string, TimeSpan>();
-        var sort = new Dictionary<int, string>();
-        var num = 0;
-        var allTimer = TimeSpan.Zero;
-
-        for (int i = Order.Count - 1; i >= 0; i--)
-        {
-            foreach (var system in Dic[Order[i]])
-            {
-                var alltimelong = watch.Elapsed;
-                try
-                {
-                    watch.Start();
-                    await system.Item2.UnInstall();
-                    watch.Stop();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine("Automation UnInstallation System : {0} -> Error : {1}", system.Item1.Name, ex);
-                    watch.Stop();
-                }
-
-                var timer = watch.Elapsed - alltimelong;
-                var title = string.Format("{0} {1}", num, system.Item1.Name);
-                allDic.Add(title, timer);
-                sort.Add(num, title);
-                allTimer += timer;
-                num++;
-            }
-        }
-
-        Console.WriteLine("Automation UnInstallation System Time Total -> <color=#E47833><b>{0:G}</b></color>", allTimer);
-        for (int i = 0; i < num; i++)
-        {
-            string key = sort[i];
-            TimeSpan timer = allDic[key];
-            var timeinfo = string.Format("[<color=#E47833><b>{1}% | {0:G}</b></color>]", timer, (timer.Ticks / (double)allTimer.Ticks * 100).ToString("00.0000")).PadRight(30);
-            Console.WriteLine("Automation UnInstallation System Time : {0} -> {1}", timeinfo, key);
-        }
-
-    }
-
-    #endregion
 }